License change: AGPL 3 → PostgreSQL license
I thought it better and more becoming to stick with the license that
characterizes the PostgreSQL ecosystem.  And, as long as I'm still the
only author of this extension, a license change is easily made.  So if
there was any time to stop sitting the fence on this one it was now.